listo horario para 1 solo

+++ app/Http/Controllers/horarioController.php
@@ -478,8 +478,18 @@ class horarioController extends Controller
     public function eliminarHora(Request $request)
     {
         $idHora = $request->idHora;
+        $tipoEliminar = $request->tipoEliminar;
 
-        $temporal_evento = temporal_eventos::where('id', '=', $idHora)->delete();
+        if($tipoEliminar==0){
+            $temporal_evento = temporal_eventos::where('id', '=', $idHora)->delete();
+        }
+        else{
+            $horario_Empleado=horario_empleado::findOrfail($idHora);
+            $horario_Empleado->estado=0;
+            $horario_Empleado->save();
+
+        }
+
     }
     public function cambiarEstado(Request $request)
     {
@@ -1487,7 +1497,7 @@ class horarioController extends Controller
 
         $incidencias = DB::table('incidencias as i')
             ->select([
-                'idi.inciden_dias_id as id', 'i.inciden_descripcion as title', 'i.inciden_descuento as color', 'i.inciden_descuento as textColor',
+                'idi.inciden_dias_id as id', 'i.inciden_descripcion as title', 'i.inciden_descuento as color', 'i.inciden_descripcion as textColor',
                 'idi.inciden_dias_fechaI as start', 'idi.inciden_dias_fechaF as end', 'i.inciden_descripcion as horaI', 'i.inciden_descripcion as horaF', 'i.inciden_descripcion as borderColor', 'laborable',
                 'i.inciden_descripcion as horaAdic', 'i.inciden_descripcion as idhorario', 'i.inciden_descripcion as horasObliga', 'i.inciden_descripcion as nHoraAdic',
             ])
